- Melhorando organização do código - Agora só passa pro próximo jogador após uma jogada valida
