+++ primeiro-projeto/src/Tipos.hs
@@ -1,0 +1,16 @@
+module Tipos where
+
+{-
+    Definição de tipos e constantes
+-}
+
+type Tabuleiro = [[Char]]
+type Coord = (Int, Int)
+
+vazio, jogadorX, jogadorO :: Char
+vazio = ' '
+jogadorX = 'X'
+jogadorO = 'O'
+
+tamanhoTabuleiro :: Int
+tamanhoTabuleiro = 3
